chore: Rename `meshing_two` module to `meshing`
