+++ src/renderer.rs
@@ -3,7 +3,7 @@ use bevy::{
     window::{CursorGrabMode, PrimaryWindow},
 };
 
-pub mod meshing_two;
+pub mod meshing;
 pub mod skybox;
 pub struct MacawRendererPlugin;
 
@@ -16,8 +16,8 @@ impl Plugin for MacawRendererPlugin {
 impl MacawRendererPlugin {
     fn setup(
         mut commands: Commands,
-        mut meshes: ResMut<Assets<Mesh>>,
-        mut materials: ResMut<Assets<StandardMaterial>>,
+        meshes: ResMut<Assets<Mesh>>,
+        materials: ResMut<Assets<StandardMaterial>>,
         mut window_query: Query<&mut Window, With<PrimaryWindow>>,
         asset_server: Res<AssetServer>,
     ) {
@@ -25,29 +25,8 @@ impl MacawRendererPlugin {
         let mut window = window_query.single_mut();
         window.cursor.grab_mode = CursorGrabMode::Locked;
 
-        let world = crate::world::World::generate_test_chunk();
+        let world = crate::world::generation::Generate::testing_world();
 
-        let mesh = Cuboid::new(1.0, 1.0, 1.0).mesh();
-
-        commands.spawn(PbrBundle {
-            mesh: meshes.add(mesh),
-            transform: Transform {
-                translation: Vec3::new(100.0, 0.0, 0.0),
-                ..Default::default()
-            },
-            // material: `get_texture_from_block_type(block_type)`,
-            material: materials.add(StandardMaterial {
-                base_color: Color::RED,
-                base_color_texture: Some(
-                    asset_server.load("/home/barrett/Documents/macaw/assets/stone.png"),
-                ),
-                reflectance: 1.0,
-                metallic: 0.1,
-                ..Default::default()
-            }),
-            ..Default::default()
-        });
-
-        meshing_two::render_clusters(&mut commands, meshes, world, asset_server, materials);
+        meshing::render_clusters(&mut commands, meshes, world, asset_server, materials);
     }
 }
